Make block template abstract and protected add block methods
